Removed exposure of helper functions
