+++ lib/bedrock/index.js
@@ -1,26 +1,13 @@
-const buildIndexFromArray = require('../indexer')
-
 module.exports = (data) => {
   return {
-
     handleStartGame (packet) {
-      this.loadItemStates(packet.itemstates)
+      const { loadItemStates, loadRuntimeIds, loadHashedRuntimeIds } = require('./loader')(this, data)
+      loadItemStates(packet.itemstates)
       if (this.supportFeature('blockHashes') && packet.block_network_ids_are_hashes) {
-        this.loadHashedRuntimeIds()
+        loadHashedRuntimeIds()
       } else {
-        this.loadRuntimeIds()
+        loadRuntimeIds()
       }
-    },
-
-    loadItemStates (itemStates) {
-      const items = []
-      for (const item of itemStates) {
-        const name = item.name.replace('minecraft:', '')
-        items.push({ ...data.itemsByName[name], name, id: item.runtime_id })
-      }
-      data.itemsArray = items
-      data.items = buildIndexFromArray(data.itemsArray, 'id')
-      data.itemsByName = buildIndexFromArray(data.itemsArray, 'name')
     },
 
     writeItemStates () {
@@ -37,23 +24,6 @@ module.exports = (data) => {
       }
 
       return itemstates
-    },
-
-    loadHashedRuntimeIds () {
-      const Block = require('prismarine-block')(this)
-      data.blocksByRuntimeId = {}
-      for (let i = 0; i < data.blockStates.length; i++) {
-        const { name, states } = data.blockStates[i]
-        const hash = Block.getHash(name, states)
-        data.blocksByRuntimeId[hash] = { stateId: i, ...data.blocksByStateId[i] }
-      }
-    },
-
-    loadRuntimeIds () {
-      data.blocksByRuntimeId = {}
-      for (let i = 0; i < data.blockStates.length; i++) {
-        data.blocksByRuntimeId[i] = { stateId: i, ...data.blocksByStateId[i] }
-      }
     }
   }
 }
